# memory/ag_export/NIVXRAY_COMPLETE_AG_EXPORT/01_COMPLETE_SOURCE/backend/baselines/openapi_snapshot.json
{
  "openapi": "3.1.0",
  "paths": {
...
    "/api/cases": {
      "get": {
        "has_request_body": false,
        "responses": [
          "200",
          "422"
        ]
      }
    },
...
    "/api/cases/{case_id}": {
      "delete": {
        "has_request_body": false,
        "responses": [
          "200",
          "422"
        ]
      },
      "get": {
        "has_request_body": false,
        "responses": [
          "200",
          "422"
        ]
      }
    },
...
  }
}